Support embedded files that are non-bitcode.
Files created with Apple's version of LLVM can contain more than just
bitcode in the embedded xar archive. This commit adds support for
recursively nested xar's and embedded Mach-O's. The latter appear when
non-inline assembly is used.

Other improvements included in this commit is support for address
sanitizer (asan), vendoring of the Catch unit test framework, the
addition of a CMake-generated version and the use of UUIDs for the
extracted files.